backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-extra-fields==3.7.0
psycopg2-binary==2.9.10
pillow==11.0.0
sqlparse==0.5.1
urllib3==2.2.3
